fixes
feedback #9: use fstat to get the file size. I'm disliking that I have
to cast to size_t when creating a HttpEntity that has a file as its
content. I'm reluctant to introduce two different HttpEntity classes for
now (like ContentHttpEntity and FileHttpEntity). Maybe in the future.
But I'd like to hear your thoughts.